requirements.txt
python-telegram-bot >= 20.3
requests >= 2.30.0
beautifulsoup4 >= 4.12.2
Django==4.1.3